+++ docs/html/search/all_c.js
@@ -8,8 +8,12 @@ var searchData=
   ['pin_5fopen_5fdrain_5finit_5',['pin_open_drain_init',['../group__pin.html#gaa910c8f32cf6f9311e1cba9490d14e1e',1,'pin.h']]],
   ['pin_5fpush_5fpull_5finit_6',['pin_push_pull_init',['../group__pin.html#ga2c7e5cb2a19518afffd55e46270f2e41',1,'pin.h']]],
   ['pin_5fquasi_5fbidiretional_5finit_7',['pin_quasi_bidiretional_init',['../group__pin.html#ga76365dbd738481635f5d73183e2059d3',1,'pin.h']]],
-  ['power_20management_8',['Power management',['../group__power__management.html',1,'']]],
-  ['power_5fdown_9',['power_down',['../group__power__management.html#ga1e0d38ad57d2078db65a84e857f5e5b6',1,'power_management.h']]],
-  ['power_5fmanagement_2eh_10',['power_management.h',['../power__management_8h.html',1,'']]],
-  ['putchar_11',['putchar',['../group__uart.html#ga07036e93939040c563e862a6a31b65bc',1,'putchar(int ch):&#160;uart.c'],['../group__uart.html#ga07036e93939040c563e862a6a31b65bc',1,'putchar(int ch):&#160;uart.c']]]
+  ['port_5fmode_5finput_5fonly_8',['port_mode_input_only',['../group__pin.html#ga87089fcfb4370b925ccff27cd7a2d5d7',1,'pin.h']]],
+  ['port_5fmode_5fopen_5fdrain_9',['port_mode_open_drain',['../group__pin.html#ga2537c3e635cf8b2900be83638db9f026',1,'pin.h']]],
+  ['port_5fmode_5fpull_5fpush_10',['port_mode_pull_push',['../group__pin.html#ga4aa18d16ac59c0e3fac513e5a0d3b0ed',1,'pin.h']]],
+  ['port_5fmode_5fquasi_5fbidiretional_11',['port_mode_quasi_bidiretional',['../group__pin.html#gabcdf06a6d669c9b45515672ae7bb52d3',1,'pin.h']]],
+  ['power_20management_12',['Power management',['../group__power__management.html',1,'']]],
+  ['power_5fdown_13',['power_down',['../group__power__management.html#ga1e0d38ad57d2078db65a84e857f5e5b6',1,'power_management.h']]],
+  ['power_5fmanagement_2eh_14',['power_management.h',['../power__management_8h.html',1,'']]],
+  ['putchar_15',['putchar',['../group__uart.html#ga07036e93939040c563e862a6a31b65bc',1,'putchar(int ch):&#160;uart.c'],['../group__uart.html#ga07036e93939040c563e862a6a31b65bc',1,'putchar(int ch):&#160;uart.c']]]
 ];