+++ include/timer0_mode0.h
@@ -146,6 +146,21 @@
 }
 
 /**
+ * @brief Reload timer0 mode0 timer TL0 and TH0 on the fly
+ * @details Reload is available in modes 0, 2, 3.
+ * 
+ * @param th0 uint8_t TH0 value
+ * @param tl0 uint8_t TL0 value
+ * 
+ * @ingroup timer0_mode0
+ */
+#define timer0_mode0_direct_reload(th0, tl0)                \
+{                                                           \
+    TL0 = tl0;                                              \
+    TH0 = th0;                                              \
+}
+
+/**
  * @brief Run timer0.
  * @details Before run timer0_mode0_12T_init or timer0_mode0_1T_init should be called.
  *