+++ compile_examples.py
@@ -1,7 +1,10 @@
 import os
 
+from SCons.Script import SConscript
+
 Import("env")
 
 print("Add examples to build configuration")
 
-env.Execute(['cd examples ; scons ; rm -rf build'])
+env.SConscript('examples/SConstruct', exports = 'env')
+#env.Execute(['cd examples ; scons ; rm -rf build'])