+++ compile_examples.py
@@ -4,7 +4,8 @@ from SCons.Script import SConscript
 
 Import("env")
 
-print("Add examples to build configuration")
+print("Building examples")
 
 env.SConscript('examples/SConstruct', exports = 'env')
-#env.Execute(['cd examples ; scons ; rm -rf build'])
+
+print("Build examples finished")