+++ docs/html/search/all_1.js
@@ -1,7 +1,6 @@
 var searchData=
 [
-  ['chip_5fid_2eh_0',['chip_id.h',['../chip__id_8h.html',1,'']]],
-  ['chip_5fid_5faddr_5from_1',['CHIP_ID_ADDR_ROM',['../group__chip__id.html#gade5a33c9ca0579ba754154984706deff',1,'chip_id.h']]],
-  ['chip_5fid_5fsize_2',['CHIP_ID_SIZE',['../group__chip__id.html#ga909ec39ac520a4448627b2d9aa61d016',1,'chip_id.h']]],
-  ['chipid_3',['ChipID',['../group__chip__id.html',1,'']]]
+  ['add_20library_20to_20your_20project_0',['Add library to your project',['../index.html#autotoc_md2',1,'']]],
+  ['aux_1',['aux',['../group__timer0__to__ms.html',1,'Timer0 Aux'],['../group__timer2__to__ms.html',1,'Timer2 Aux']]],
+  ['aux_20common_2',['Timer Aux common',['../group__timer__to__ms.html',1,'']]]
 ];