+++ docs/html/search/all_2.js
@@ -1,19 +1,8 @@
 var searchData=
 [
-  ['delay_0',['Delay',['../group__delays.html',1,'']]],
-  ['delay_2eh_1',['delay.h',['../delay_8h.html',1,'']]],
-  ['delay_5fms_2',['delay_ms',['../group__delays.html#ga389926e6490f1d3db0f0896e8b0f5f2e',1,'delay.h']]],
-  ['disable_5fadc_5finterrupt_3',['disable_adc_interrupt',['../group__interrupts.html#ga499a9a838de2728a9c6ab4311d055b86',1,'interrupt.h']]],
-  ['disable_5fint0_5finterrupt_4',['disable_int0_interrupt',['../group__interrupts.html#ga5280b5134c4d1fad2f492e30205d7691',1,'interrupt.h']]],
-  ['disable_5fint1_5finterrupt_5',['disable_int1_interrupt',['../group__interrupts.html#gaa588f32201651fc565dd745d36048a97',1,'interrupt.h']]],
-  ['disable_5fint2_5finterrupt_6',['disable_int2_interrupt',['../group__interrupts.html#ga4668129f676d8d123d5992da8b7d6564',1,'interrupt.h']]],
-  ['disable_5fint3_5finterrupt_7',['disable_int3_interrupt',['../group__interrupts.html#ga0eefb72a54f3befd3535d70e71a06de4',1,'interrupt.h']]],
-  ['disable_5fint4_5finterrupt_8',['disable_int4_interrupt',['../group__interrupts.html#gae3a80148c904b86a58976a26399720d0',1,'interrupt.h']]],
-  ['disable_5flow_5fvoltage_5finterrupt_9',['disable_low_voltage_interrupt',['../group__interrupts.html#gae7bd7f894582b7f376b077fbd4435864',1,'interrupt.h']]],
-  ['disable_5fmaster_5fclock_5foutput_10',['disable_master_clock_output',['../group__freq.html#gabcec59e26510c1132c4f13f45b13ec6b',1,'frequency.h']]],
-  ['disable_5fmcu_5finterrupts_11',['disable_mcu_interrupts',['../group__interrupts.html#ga8b0ea5804af6316376f49ade70b868b3',1,'interrupt.h']]],
-  ['disable_5fspi_5finterrupt_12',['disable_spi_interrupt',['../group__interrupts.html#gad8ec33652ec7e3a2a2754678217a9799',1,'interrupt.h']]],
-  ['disable_5ftimer0_5finterrupt_13',['disable_timer0_interrupt',['../group__interrupts.html#ga2ef6ebd3211c16ce240018d22589e8da',1,'interrupt.h']]],
-  ['disable_5ftimer2_5finterrupt_14',['disable_timer2_interrupt',['../group__interrupts.html#gac8daf51c922df876c57f419b63996dd5',1,'interrupt.h']]],
-  ['disable_5fuart1_5finterrupt_15',['disable_uart1_interrupt',['../group__interrupts.html#ga5dac79f9182e6dc1a31448a28ee0a162',1,'interrupt.h']]]
+  ['chip_5fid_2eh_0',['chip_id.h',['../chip__id_8h.html',1,'']]],
+  ['chip_5fid_5faddr_5from_1',['CHIP_ID_ADDR_ROM',['../group__chip__id.html#gade5a33c9ca0579ba754154984706deff',1,'chip_id.h']]],
+  ['chip_5fid_5fsize_2',['CHIP_ID_SIZE',['../group__chip__id.html#ga909ec39ac520a4448627b2d9aa61d016',1,'chip_id.h']]],
+  ['chipid_3',['ChipID',['../group__chip__id.html',1,'']]],
+  ['common_4',['common',['../group__timer__to__ms.html',1,'Timer Aux common'],['../group__timer__common.html',1,'Timer common']]]
 ];