+++ docs/html/search/all_9.js
@@ -1,7 +1,9 @@
 var searchData=
 [
-  ['main_5ffosc_0',['MAIN_Fosc',['../sys_8h.html#aa52cf4efd75d43c7917d9eec3becd2fa',1,'sys.h']]],
-  ['main_5ffosch_1',['MAIN_FoscH',['../sys_8h.html#ae1b9b7f30c88b5491fd35d286418f6d6',1,'sys.h']]],
-  ['main_5ffoscl_2',['MAIN_FoscL',['../sys_8h.html#ab42e14bee431f5da3ecef889588cc728',1,'sys.h']]],
-  ['master_5fclock_5foutput_5fpin_5ft_3',['master_clock_output_pin_t',['../group__freq.html#gae5d505beabdbd466d648cb5dccf4b6d2',1,'frequency.h']]]
+  ['level_20stc15_20prorgamming_0',['Low level STC15 prorgamming',['../index.html#autotoc_md3',1,'']]],
+  ['library_20documentation_1',['Library documentation',['../index.html#autotoc_md6',1,'']]],
+  ['library_20modules_20maturity_2',['Library modules maturity',['../index.html#autotoc_md5',1,'']]],
+  ['library_20to_20your_20project_3',['Add library to your project',['../index.html#autotoc_md2',1,'']]],
+  ['low_4',['LOW',['../group__interrupts.html#gga9ebbbf91c4e6fa77cced0ac0daefd964a6a226f4143ca3b18999551694cdb72a8',1,'interrupt.h']]],
+  ['low_20level_20stc15_20prorgamming_5',['Low level STC15 prorgamming',['../index.html#autotoc_md3',1,'']]]
 ];