+++ docs/html/search/groups_0.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['chipid_0',['ChipID',['../group__chip__id.html',1,'']]]
+  ['aux_0',['aux',['../group__timer0__to__ms.html',1,'Timer0 Aux'],['../group__timer2__to__ms.html',1,'Timer2 Aux']]],
+  ['aux_20common_1',['Timer Aux common',['../group__timer__to__ms.html',1,'']]]
 ];