+++ docs/html/search/groups_4.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['pin_0',['Pin',['../group__pin.html',1,'']]],
-  ['power_20management_1',['Power management',['../group__power__management.html',1,'']]]
+  ['interrupt_0',['Interrupt',['../group__interrupts.html',1,'']]]
 ];