+++ docs/html/search/groups_6.js
@@ -1,12 +1,5 @@
 var searchData=
 [
-  ['timer_20aux_20common_0',['Timer Aux common',['../group__timer__to__ms.html',1,'']]],
-  ['timer_20common_1',['Timer common',['../group__timer__common.html',1,'']]],
-  ['timer0_20aux_2',['Timer0 Aux',['../group__timer0__to__ms.html',1,'']]],
-  ['timer0_20mode0_3',['Timer0 Mode0',['../group__timer0__mode0.html',1,'']]],
-  ['timer0_20mode1_4',['Timer0 Mode1',['../group__timer0__mode1.html',1,'']]],
-  ['timer0_20mode2_5',['Timer0 Mode2',['../group__timer0__mode2.html',1,'']]],
-  ['timer0_20mode3_6',['Timer0 Mode3',['../group__timer0__mode3.html',1,'']]],
-  ['timer2_20aux_7',['Timer2 Aux',['../group__timer2__to__ms.html',1,'']]],
-  ['timer2_20mode0_8',['Timer2 Mode0',['../group__timer2__mode0.html',1,'']]]
+  ['pin_0',['Pin',['../group__pin.html',1,'']]],
+  ['power_20management_1',['Power management',['../group__power__management.html',1,'']]]
 ];