+++ docs/html/search/groups_7.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['uart1_0',['UART1',['../group__uart.html',1,'']]]
+  ['reset_0',['Reset',['../group__mcu__reset.html',1,'']]]
 ];