+++ docs/html/search/groups_9.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['uart1_0',['UART1',['../group__uart.html',1,'']]]
+];