+++ docs/html/search/pages_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['version_20history_0',['Version history',['../md_VERSION_HISTORY.html',1,'']]]
+  ['history_0',['Version history',['../md_VERSION__HISTORY.html',1,'']]]
 ];