+++ docs/html/search/pages_1.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['what_20it_20is_0',['What it is',['../index.html',1,'']]]
+  ['is_0',['What it is',['../index.html',1,'']]],
+  ['it_20is_1',['What it is',['../index.html',1,'']]]
 ];