+++ docs/html/search/pages_2.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['version_20history_0',['Version history',['../md_VERSION__HISTORY.html',1,'']]]
+];