+++ VERSION_HISTORY.md
@@ -77,7 +77,7 @@
 * Pin module set port mode routines implemented
 
 Next releases todo
-* TODO Implement timer_modeX_direct_init(TxH, TxL) for direct set TLx, THx registers
+* TODO Implement timer_modeX_direct_*(TxH, TxL) for direct set TLx, THx registers
 * TODO Does not recompile examples every time. Compile only when it changed.
 * TODO Suppress compiler output for examples on non verbose build 
 * TODO Implement counters