+++ VERSION_HISTORY.md
@@ -77,6 +77,8 @@
 * Pin module set port mode routines implemented
 
 Next releases todo
+* Fix timer0_mode2_reload - TH0 value should be set, now used TL0
+* Fix examples/mode0/timer2 compilation
 * TODO Implement timer_modeX_direct_*(TxH, TxL) for direct set TLx, THx registers
 * TODO Does not recompile examples every time. Compile only when it changed.
 * TODO Suppress compiler output for examples on non verbose build