+++ VERSION_HISTORY.md
@@ -75,11 +75,11 @@
 
 0.9
 * Pin module set port mode routines implemented
+* Fix examples/mode0/timer2 compilation
+* Fix timer0_mode2_reload - TH0 value should be set, now used TL0
 
 Next releases todo
 * TODO Implement timer_modeX_direct_*(TxH, TxL) for direct set TLx, THx registers
-* Fix timer0_mode2_reload - TH0 value should be set, now used TL0
-* Fix examples/mode0/timer2 compilation
 * TODO Does not recompile examples every time. Compile only when it changed.
 * TODO Suppress compiler output for examples on non verbose build 
 * TODO Implement counters