+++ VERSION_HISTORY.md
@@ -79,10 +79,10 @@
 * Fix timer0_mode2_reload - TH0 value should be set, now used TL0
 * Fix INT0 should be defined in timer0 modes
 * timerX_modeX_direct_{reload, start}(TxH, TxL) for direct set TLx, THx registers implemented
-* Does not recompile examples every time. Compile only when it changed.
+* Does not recompile examples every time. Compile only when it changed
+* Compiler output suppressed for examples building in non verbose mode 
 
 Next releases todo
-* TODO Suppress compiler output for examples on non verbose build 
 * TODO Implement counters
 * TODO Implement UART
 * TODO Optimize timer2 baud rate calculation. Use precalculated array of baudrates